+++ packages/interact/src/core/css.ts
@@ -12,6 +12,7 @@ import type {
 import {
   kebabCustomProp,
   camelToKebabCase,
+  getStateStyleProperties,
   transitionEffectToTransitionsList,
   getFullPredicateByType,
   getSelectorCondition,
@@ -268,7 +269,7 @@ function effectToCSS(
   } else if (transition || transitionProperties) {
     usedProperties = ['transition'];
 
-    const properties = transition?.styleProperties || transitionProperties || [];
+    const properties = getStateStyleProperties(effect);
     const transitions = transitionEffectToTransitionsList(effect);
 
     // declaring transition custom property